Extra Credit Documentation Fix
I found an error in the ToDo API documentation that I want to push. The error is in the Javascript example for the GET a ToDo section: as it currently stands, the code returns an error stating that the data object referenced in the xhttp2.send() method is undefined. Additionally, the xhttp2.open() method has a "POST" parameter which adds to the array of ToDos rather than retrieves the ToDos.

I've made a few fixes in the example code in the documentation to correct these errors:
- deleted the parameter referenced in the xhttp2.send() method
- changed the "POST" parameter to a "GET" parameter in the xhttp2.open() method.